backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
motor>=3.3.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pytest>=7.4.0
httpx>=0.24.1
